tableau_clone_streamlit/requirements.txt
streamlit
pandas
plotly
requests
Pillow
openpyxl  # For Excel file support in pandas
